Add API skeleton
Let's have a data type for a 128-bit block of data.

Let's have a data type for that 4x4 AES state grid.

Let's have API for encrypting one block. It's a stub for now and making
it work is basically the entire project.

Once it is working, we have to add CBC on top, but that should be
straightforward.