Test that our encryption matches nettle
There's only one right or wrong answer. If we don't give the same result
as an established cryptographic library, then we're broken.

Fortunately, we do always give the same result as nettle, even when
testing with 10000 random inputs and keys.

This provides the basis for testing that we have implemented CBC
correctly.