gui: Remove a couple unused function parameters
I normally prefer to have the parameter list match the actual callback
declaration as it's weird to pass more arguments to a function than it
actually consumes, but nowhere in this file do we actually need the user
data parameters as I'm just putting everything in file scope, so I
omitted all of those. It's silly to omit some unused parameters but not
others, so get rid of them.